findByModuleCode CountByModuleCode

+++ easy-common/easy-rss/src/main/java/com/rongyi/rss/content/ROAContentService.java
@@ -395,5 +395,5 @@ public interface ROAContentService {
      * @author wangjh7
      * @return
      */
-    public int getCountContentByModuleCode(ContentParam contentParam) ;
+    public Long getCountContentByModuleCode(ContentParam contentParam) ;
 }
